Move up and split sub-namespace State
