+++ src/Extension/CustomStateFoundation.php
@@ -4,8 +4,8 @@
 namespace Wirecard\Order\State\Extension;
 
 
-use Wirecard\Order\State\Implementation\State\CalculableState;
-use Wirecard\Order\State\Implementation\State\StateHelper;
+use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
 use Wirecard\Order\State\State;
