+++ src/Implementation/CalculableState.php
@@ -1,9 +1,10 @@
 <?php
 
 
-namespace Wirecard\Order\State\Implementation\State;
+namespace Wirecard\Order\State\Implementation;
 
 
+use Wirecard\Order\State\Implementation\StateTransitions;
 use Wirecard\Order\State\State;
 
 interface CalculableState extends StateTransitions, State
