+++ src/Implementation/StateTransitions.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace Wirecard\Order\State\Implementation\State;
+namespace Wirecard\Order\State\Implementation;
 
 use Wirecard\Order\State\Implementation\TransitionData;
 use Wirecard\Order\State\State;
