+++ src/State/Authorized.php
@@ -1,8 +1,10 @@
 <?php
 
 
-namespace Wirecard\Order\State\Implementation\State;
+namespace Wirecard\Order\State\State;
 
+use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
 use Wirecard\Order\State\State;
