+++ src/State/Pending.php
@@ -1,11 +1,15 @@
 <?php
 
 
-namespace Wirecard\Order\State\Implementation\State;
+namespace Wirecard\Order\State\State;
 
+use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Implementation\StateHelper;
+use Wirecard\Order\State\State\Success;
 use Wirecard\Order\State\Implementation\TransitionData;
+use Wirecard\Order\State\State\Failed;
 
-class Pending implements CalculableState
+class Pending implements \Wirecard\Order\State\Implementation\CalculableState
 {
     use StateHelper;
