+++ src/State/Processing.php
@@ -1,9 +1,12 @@
 <?php
 
 
-namespace Wirecard\Order\State\Implementation\State;
+namespace Wirecard\Order\State\State;
 
+use Wirecard\Order\State\Implementation\CalculableState;
+use Wirecard\Order\State\Implementation\StateHelper;
 use Wirecard\Order\State\Implementation\TransitionData;
+use Wirecard\Order\State\State\Failed;
 
 /**
  * Class Processing
@@ -15,7 +18,7 @@ use Wirecard\Order\State\Implementation\TransitionData;
  */
 class Processing implements CalculableState
 {
-    use StateHelper;
+    use \Wirecard\Order\State\Implementation\StateHelper;
 
     public function getPossibleNextStates()
     {
