+++ tests/integration/FacadeTest.php
@@ -3,12 +3,12 @@
 use Wirecard\Order\State\CreditCardTransactionType\AuthorizationTransaction;
 use Wirecard\Order\State\CreditCardTransactionType;
 use Wirecard\Order\State\CreditCardTransactionType\PurchaseTransaction;
-use Wirecard\Order\State\Implementation\State\Authorized;
-use Wirecard\Order\State\Implementation\State\Failed;
-use Wirecard\Order\State\Implementation\State\Pending;
-use Wirecard\Order\State\Implementation\State\Processing;
-use Wirecard\Order\State\Implementation\State\Started;
-use Wirecard\Order\State\Implementation\State\Success;
+use Wirecard\Order\State\State\Authorized;
+use Wirecard\Order\State\State\Failed;
+use Wirecard\Order\State\State\Pending;
+use Wirecard\Order\State\State\Processing;
+use Wirecard\Order\State\State\Started;
+use Wirecard\Order\State\State\Success;
 use Wirecard\Order\State\OrderState;
 use Wirecard\Order\State\State;
 use Wirecard\Order\State\TransactionType;
